continue with refactoring classes
